[mapleall]Internal commit msg:     fix ivopt bugs     [mapleall] PClint warnings clean: implicit conversion 2

+++ src/mapleall/maple_be/include/cg/datainfo.h
@@ -170,7 +170,7 @@ class DataInfo {
       uint32 baseWord = 0;
       bool firstTime = true;
       while (word) {
-        int32 index = __builtin_ffsll(word);
+        int32 index = __builtin_ffsll(static_cast<int64>(word));
         if (index == 0) {
           continue;
         }
